Some sounds
I'm not sure what you had in mind for the bubble sound but see if this
one works. Let me know if you need something that can be looped.

For the pressStart sound that im including here i was thinking maybe you
can use it for the main menu for when you "Press Start" if you know what
i mean.

I'll make a different one for for misc menu clicking etc..